% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{AlignmentSTAR}
\alias{AlignmentSTAR}
\title{Use the STAR tool to align the reads}
\usage{
AlignmentSTAR(parallel_threads = 1, star_threads = 6, paired = TRUE,
  paired_pattern = c("_1.fastq", "_2.fastq"), fastqPath = "")
}
\arguments{
\item{parallel_threads}{specify how many parallel processes to spawn}

\item{star_threads}{specify how many threads star should use.}

\item{paired}{specify whether you have paried reads or not.}

\item{paired_pattern}{specify the suffix of the paried-end fastq file names.}

\item{fastqPath}{specify path to FASTQ files if different than default}
}
\description{
Use the STAR tool to align reads and generate transcriptome bam files
}
\details{
Uses STAR to align reads in FASTQ files against the reference genome. Optionally you can specify paired end reads. The code assumes
paired reads have fastq files that differ by one character (i.e. sampleA_read1.fastq, sampleA_read2.fastq) and will perform
matching of paired fastq files based on that assumption using string edit distance. Read 1 is assumed to be upstream
and read 2 is assumed to be downstream.
The number of parallel_threads*star_threads should not be more than the number of cores available on your system.
}
