% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{annotateUCSC}
\alias{annotateUCSC}
\title{Annotate features: map UCSC gene cluster id to gene symbols}
\usage{
annotateUCSC(genome = "hg38", force = TRUE)
}
\arguments{
\item{genome}{/code{character} specify the genome and version of annotation}

\item{force}{/code{logical} if TRUE force the annoation even if the feature file already exists}
}
\description{
Annotate features by mapping UCSC gene cluster ids to gene symbols.
}

