% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{annotationsFromSRX}
\alias{annotationsFromSRX}
\title{Quick and dirty annotations from SRAmetadb}
\usage{
annotationsFromSRX(x)
}
\arguments{
\item{x}{\code{character} vector of SRX accessions}
}
\description{
Gets annotations for SRR files based on SRAmetadb contents
}

