% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{assignConfig}
\alias{assignConfig}
\title{Assign a new element to the RNASeqPipelineR configuration}
\usage{
assignConfig(name, value)
}
\arguments{
\item{name}{\code{character} assign element with this name}

\item{value}{\code{value} an R object that assigned to the 'rnaseqpipeliner_configuration' \code{list} with name 'name'}
}
\description{
Assign value to name in `rnaseqpipeliner_configuration` object
}
\details{
Convenience function to assign a new or reassign an element named \code{name} to the \code{rnaseqpipeliner_configuration} object with the value \code{value}
}

