% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{buildGenomeIndexSTAR}
\alias{buildGenomeIndexSTAR}
\title{Build a genome index for STAR}
\usage{
buildGenomeIndexSTAR(path = NULL, gtf_file = "", fasta_file = NULL,
  star_threads = 1, name = NULL)
}
\arguments{
\item{path}{\code{character} specifying an \emph{absolute path} path to the Utils directory.}

\item{gtf_file}{\code{character} the name of the gtf file.}

\item{fasta_file}{\code{character} the name of the fasta file, must be specified}

\item{star_threads}{\code{integer} specify how many threads star should use.}

\item{name}{\code{character} the name of the genome output}
}
\description{
Builds a genome Index at `Utils/Reference_Genome/STARIndex`
}
\details{
You must specify the Utils path if it is not already defined, and have your genome and annotation file in a folder titled
`Reference_Genome`. This function will construct the genome index using STAR tools.
The command line is the default shown in the documentation.
A fasta_file must be provided.
gtf_file is highly recommended, and if not provided, STAR will run without annotations.
}

