% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{buildIndex}
\alias{buildIndex}
\title{Build a index for reference transcriptome}
\usage{
buildIndex(path = NULL, transcripts_fasta_file = NULL, name = NULL)
}
\arguments{
\item{path}{\code{character} specifying an \emph{absolute path} path to the Utils directory.}

\item{transcripts_fasta_file}{\code{character} the name of the transcript fasta file, must be specified.
We used the transcripts.fa file generated by RSEM rsem-prepare-reference step.}

\item{name}{\code{character} the name of the genome output.}
}
\description{
Builds a index reference transcriptome at `Utils/Reference_Genome`
}
\details{
You must specify the Utils path if it is not already defined, and have your genome in a folder titled
`Reference_Genome`. This function will construct the index reference transcriptome using kallisto tools.
The command line is the default shown in the documentation.
}

