% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildReference}
\alias{buildReference}
\title{Build a reference genome}
\usage{
buildReference(path = NULL, gtf_file = "", fasta_file = NULL,
  name = NULL)
}
\arguments{
\item{path}{\code{character} specify the path to the Utils directory.}

\item{gtf_file}{\code{character} the name of the gtf file. Empty by default}

\item{fasta_file}{\code{character} the name of the fasta file, must be specified}

\item{name}{\code{character} the name of the genome output.}
}
\description{
Builds a reference genome at `Utils/Reference_Genome`
}
\details{
You must specify the Utils path if it is not already defined, and have your genome in a folder titled
`Reference_Genome`. This function will construct the reference genome.
}

