% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{concatenateFastq}
\alias{concatenateFastq}
\title{Concatenate FastQ files}
\usage{
concatenateFastq(infile, outfile, pattern)
}
\arguments{
\item{infile}{\code{character} specify the path to the individual fastq.gz directory.}

\item{outfile}{\code{character} specify the path to the FASTQ directory}

\item{pattern}{\code{character} the string want to remove in the fastq.gz files name}
}
\description{
Combine several fastq.gz files into one fastq.gz file for each library
Decompress the fastq.gz
Copy the fastQ files into FASTQ folder
}

