% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{configure_project}
\alias{configure_project}
\title{Configure the project and store info in namespace}
\usage{
configure_project(subdirectories = NULL, fromDisk = FALSE)
}
\arguments{
\item{subdirectories}{a \code{character} vector of subdirectories passed from \code{createProject}, or a path to an existing project.}

\item{fromDisk}{\code{logical} specifying whether to load the configuration of an existing proejct, in which case the path to the proejct is given by \code{subdirectories} argument}
}
\description{
Configures the project based on the directories created in the project path
}
\details{
Stores the project configuration in the rnaseqpipeliner_configuration
variable in the packge namespace.
Generally, this function shouldn't be called by the user.
}

