% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{convertSRAtoFastQ}
\alias{convertSRAtoFastQ}
\title{Convert SRA files to FastQ}
\usage{
convertSRAtoFastQ(ncores = 8)
}
\arguments{
\item{ncores}{\code{integer} how many threads to use}
}
\description{
Uses CLI SRA Toolkit to convert SRA files to FastQ if they don't already exist.
}
\details{
Converts SRA files to FastQ using the SRA Toolkit. The function checks whether the FastQ files exist, and if not, performs the conversion.
}

