% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{createProject}
\alias{createProject}
\title{Create a new RNASeqPipeline project}
\usage{
createProject(project_name, path = ".", verbose = FALSE,
  load_from_immport = FALSE)
}
\arguments{
\item{project_name}{\code{character} name (ie, subdirectory) for project}

\item{path}{\code{character} Root directory for project}

\item{verbose}{\code{logical} Should verbose output be given?}

\item{load_from_immport}{\code{logical} Creates a 'Tab' directory for Immport tables if the data are loaded from Immport.}
}
\description{
Create the skeleton for a new RNASeqPipeline project.
}
\details{
createProject will create a new RNASeqPipeline 
project under directory 'project_name' in the path specified by 'path'.
The function creates the directory structure libraryd by RNASeqPipeline
within the new project directory, including locations for fastQ files, 
fastQC output, RSEM quantification, and optionally GEO and SRA files if the data
must be downloaded or linked to a GEO accession. Standard locations will also be provided 
to annotate the data utilizing the Immport schema. If the project exists, it will load the configuration 
if it can find it, otherwise it will proceed to re-configure the project.
}
\examples{
# construct a projects skeleton in a new folder titled "myproject".
createProject("myproject",path=".")
createProject("myproject",path=".",verbose=TRUE)
\dontrun{
createProject("myproject",path=".",load_from_immport=TRUE) #won't work unless immport is set up
}
}

