% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{detectAspera}
\alias{detectAspera}
\title{Detect and configure the aspera client location}
\usage{
detectAspera(path = NULL)
}
\arguments{
\item{path}{\code{character} path to the aspera client. If NULL, will try to detect the location based on the OS}
}
\description{
Detect the installation location of aspera client based on expected defaults
}
\details{
Sets the location of aspera client based on expected defaults. Or pass the installation path
}
\note{
Needs to be tested on linux and modified to locate the identity keys for aspera on linux
}
