% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{devel_truncateData}
\alias{devel_truncateData}
\title{Utility function to truncate the test data to `n` SRA files for testing}
\usage{
devel_truncateData(n = 2)
}
\arguments{
\item{n}{\code{integer} the number of files to limit}
}
\description{
This will truncate the GSM_table to `n` files for testing purposes.
}
\details{
Since there is over 100Gb of data, we truncate to n (default 2) files for testing.
}
