% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadSRA}
\alias{downloadSRA}
\title{Download FastQ files based on Immport data and SRA db, unless they already exist.}
\usage{
downloadSRA()
}
\description{
Based on the SRA db, download the FastQ files, unless they already exist.
}
\details{
Conditionally download the FastQ files in the SRA db based on Immport tables.
If the files are already present, they will not be downloaded.
}

