% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Trim_Bridge_UMI_WellID.R
\name{findBridge}
\alias{findBridge}
\title{Look for bridge adapter, UMI and wellID + TGGG.}
\usage{
findBridge(fq, adapter, adapterTol = 0, suffix = "", minQualityTol,
  minAlphaQual, wellIDTable)
}
\arguments{
\item{fq}{fastq}

\item{adapter}{\code{character} adapter sequence}

\item{adapterTol}{number of mismatches tolerated}

\item{suffix}{\code{character} expected terminating sequence}

\item{minQualityTol}{\code{integer} this number of bases below \code{minAlphaQual} will cause a putative adaptor match to be ignored}

\item{minAlphaQual}{\code{character} quality threshold as a phred \code{character}}

\item{wellIDTable}{\code{data.table} in wellIDsequences are taken}
}
\value{
\code{data.table}
}
\description{
Report matching start positions for bridge, and matching UMI and wellID.
}

