% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getAndConnectSRAdb}
\alias{getAndConnectSRAdb}
\title{Checks if the SRAdb sqlite database is available}
\usage{
getAndConnectSRAdb(path = NULL)
}
\arguments{
\item{path}{\code{character} path to SRAmetadb.sqlite file exists or will be placed}
}
\description{
If the database files is not available, it is downloaded.
}
\details{
Downloads and stores the SRAdb sqlite database if necessary and stores it in 
the Utils folder.
}
