% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getConfig}
\alias{getConfig}
\title{Get the configuration for the project from the namespace}
\usage{
getConfig()
}
\description{
Retrieve the configuration for the project from the namespace
}

