% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getConfigDir}
\alias{getConfigDir}
\title{Retrieve full path to 'dir' in project configuration structure}
\usage{
getConfigDir(dir)
}
\arguments{
\item{dir}{\code{character} directory that path is required for}
}
\value{
\code{character} full path to directory
}
\description{
Retrieve full path to 'dir' in project configuration structure. If 'dir'
not found exit with error.
}
