% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getConfigFile}
\alias{getConfigFile}
\title{Retrieve full path to 'dir/file' in project configuration structure}
\usage{
getConfigFile(dir, file)
}
\arguments{
\item{dir}{\code{character} directory containing file}

\item{file}{\code{character} file sought after}
}
\value{
\code{character} full path to directory
}
\description{
Retrieve full path to 'dir/file' in project configuration structure. If
file not found exit with error.
}

