% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getDataFromSRX}
\alias{getDataFromSRX}
\title{Download SRA files from SRX accessions}
\usage{
getDataFromSRX(x = NULL)
}
\arguments{
\item{x}{\code{character} a vector of SRX accession numbers}
}
\description{
Download SRA files from SRX accessions. Downloads asynchronously. Won't message you when complete, so can't be run in
batch mode at the moment.
}

