% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getDir}
\alias{getDir}
\title{Obtain directory path for configuration variable.}
\usage{
getDir(directory)
}
\arguments{
\item{directory}{configuration label}
}
\description{
Obtain directory path for configuration variable.
}
