% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSRAdb}
\alias{getSRAdb}
\title{Checks if the SRAdb sqlite database is available}
\usage{
getSRAdb(path = NULL)
}
\description{
If the database files is not available, it is downloaded.
}
\details{
Downloads and stores the SRAdb sqlite database if necessary and stores it in
the Utils folder.
}

