% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq_kallisto.R
\name{kallistoAssembleOutput}
\alias{kallistoAssembleOutput}
\title{Combine all count and tpm results from kallisto output into a single table. Also 
read error files and extract read count, aligned read count, and estimated average
length of reads.}
\usage{
kallistoAssembleOutput(paired = TRUE)
}
\arguments{
\item{paired}{If true then paired data else single read data. Default is
TRUE.}
}
\description{
Combine all count and tpm results from kallisto output into a single table. Also 
read error files and extract read count, aligned read count, and estimated average
length of reads.
}
