% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq_kallisto.R
\name{kallistoAssembleQC}
\alias{kallistoAssembleQC}
\title{Assemble kallisto and fastqc QC information and potentially combine with 
annotation data.}
\usage{
kallistoAssembleQC(paired = TRUE, doAnnotation = TRUE)
}
\arguments{
\item{paired}{if TRUE then paired end data else single read data}

\item{doAnnotation}{if TRUE then add annotation data to qc info. The 
annotation file must be the sole .csv file in the RAW_ANNOTATIONS directory.}
}
\description{
Assemble kallisto and fastqc QC information and potentially combine with 
annotation data.
}
