% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq_kallisto.R
\name{kallistoBuildTranscriptIndex}
\alias{kallistoBuildTranscriptIndex}
\title{kallisto methods}
\usage{
kallistoBuildTranscriptIndex(path = NULL,
  fasta_file = "Homo_sapiens.GRCh38.cdna.all.fa", name = "GRCH38_R91.idx",
  force = FALSE)
}
\arguments{
\item{path}{path to directory containing transcript fasta file}

\item{fasta_file}{name of fasta file to be used}

\item{name}{name of index to be built}

\item{force}{if TRUE build index if it already exists else don't.}
}
\description{
Build transcript reference index for kallisto
}
