% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{KALLISTOCalculateExpression}
\alias{KALLISTOCalculateExpression}
\title{Use the Kallisto tool to align the reads and calcuate the count and tpm values for each sample}
\usage{
KALLISTOCalculateExpression(parallel_threads = 1, paired = TRUE,
  nchunks = 10, slurm = FALSE, slurm_partition = "gottardo_r")
}
\arguments{
\item{parallel_threads}{\code{integer} specify how many parallel processes to spawn}

\item{paired}{\code{logical} specify whether you have paried reads or not.}

\item{nchunks}{\code{integer} number of chunks to split the files for a slurm job. Ignored if slurm = FALSE}

\item{slurm}{\code{logical} if \code{TRUE} job is submitted as a slurm batch job, otherwise it's run on the local machine. Slurm jobs will honour the nchunks and days_requested arguments.}

\item{slurm_partition}{\code{character} the slurm partition to submit to. Ignored if slurm=FALSE}
}
\description{
Uses Kallisto to align reads in FASTQ files against the reference genome. Optionally you can specify paired end reads. The code assumes
paired reads have fastq files that differ by one character (i.e. sampleA_read1.fastq, sampleA_read2.fastq) and will perform
matching of paired fastq files based on that assumption using string edit distance. Read 1 is assumed to be upstream
and read 2 is assumed to be downstream.
The number of parallel_threads should not be more than the number of cores available on your system.
}

