% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq_kallisto.R
\name{kallistoReadQC}
\alias{kallistoReadQC}
\title{Read kallisto error files and extract and write read count, aligned read count, and 
estimated average length of reads.}
\usage{
kallistoReadQC(paired = TRUE)
}
\arguments{
\item{paired}{If true then paired data else single read data. Default is
TRUE.}
}
\description{
Asemble kallisto qc data
}
