% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{loadImmportTables}
\alias{loadImmportTables}
\title{Load data from Immport tables}
\usage{
loadImmportTables(warn = -1, verbose = 0)
}
\arguments{
\item{verbose}{\code{integer} how verbose should we be about what's happening. 0 (default) is quiet.}

\item{\code{integer}}{specifies whether to print warnings. -1 suppresses warnings (default)}
}
\description{
Loads the GEO accession numbers from Immport tables
}
\details{
Load data from Immport tables in the \code{Tab} directory.
Stores the tables in the configuration in the namespace.
}

