% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadProject}
\alias{loadProject}
\title{Load an RNASeqPipelineR project}
\usage{
loadProject(project_dir = NULL, name = NULL)
}
\arguments{
\item{project_dir}{\code{character} the path to the project}

\item{name}{\code{character} the name of the project}
}
\description{
Load and RNASeqPipelineR project
}
\details{
Load an RNASeqPipelineR project from disk.
}

