% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{mergeData}
\alias{mergeData}
\title{Map the annotation file to the count and tpm data}
\usage{
mergeData(annotationMatch = NULL)
}
\arguments{
\item{annotationMatch}{\code{character} Column name of annotation file
that maps to the column names of the count and tpm data sets (fastq
file names with the suffixes removed including paired end identifiers
if they exist). Default column name used is 'Sample'.}
}
\value{
\code{list} list containing the count, tpm, feature, and
annotation data. list names are counts, tpms, featureData, annoData.

The annotation file, which contains the experimental design information,
must be constructed by the user and placed in the RAW_ANNOTATIONS
directory. It must be a .csv file and be the only .csv file in the
RAW_ANNOTATION directory. By default 'mergeData' will look for a column
called 'Sample' that maps each row of the annotation file to the columns
of the counts and tpm columns.
The quality control matrix is generated by running 'runFASTQ' and then
'QualityControl'
}
\description{
Match the count, tpm, and feature data to the annotation file. Remove
samples in count and tpm that are not found in the annotation file and
ensure that the count and tpm columns match the rows of the annoation
file. Also attach the results of the quality_control table to the
annotation file.
}

