% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MiTCR}
\alias{MiTCR}
\title{Run MiTCR on each fastQ file}
\usage{
MiTCR(gene = "TRB", species = NULL, ec = 2, pset = "flex")
}
\arguments{
\item{gene}{\code{character} c("TRB","TRA")}

\item{species}{\code{character} c("hs","mm")}

\item{ec}{\code{integer}, c(0,1,2)}

\item{pset}{\code{character} c("flex")}
}
\description{
Run the MiTCR tool on each fastQ file
}
\details{
Runs MiTCR on each fastQ file.
}

