% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Trim_Bridge_UMI_WellID.R
\name{partitionFastqFiles}
\alias{partitionFastqFiles}
\title{Use a sampleSheet to partition fastq by wellID}
\usage{
partitionFastqFiles(sampleSheet, out.dir = NULL, minQualityTol = 4,
  minQuality = 15)
}
\arguments{
\item{sampleSheet}{\code{data.table giving the map between wellID and input fastq file}}

\item{out.dir}{directory that partitioned fastq will be written.  If NULL, then will use the FASTQ directory from the project.}

\item{minQualityTol}{\code{integer} number of bases below minQuality that will be tolerated in bridge-umi-wellID-suffix}

\item{minQuality}{\code{numeric} \code{minQualityTol} or more bases below this numeric quality in the bridge-umi-wellID-suffix will result in the read being discarded}
}
\value{
Returns a \code{data.table} providing summary statistics (how many codes could be determined unambiguously by wellID and input fq1, as well as how many reads couldn't be matched by code or were ambiguous).
}
\description{
Each fastq (fq1 and fq2) will be divided by wellID.
Output file names are determined by column sampleID + _1.fastq and _2.fastq.
Samples that appear repeatedly in the sample sheet are concatenated.
Reading fastq files, and writing output is parallelized. Set \code{options(mc.cores=NCORES)} to utilize.
We can't easily parallelize over input fastq if we might be concatenating across input fastq files.
}

