% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{partitionReport}
\alias{mappingReport}
\alias{partitionReport}
\title{Generate reports on partitioning/trimming/qc}
\usage{
partitionReport()

mappingReport(...)
}
\arguments{
\item{...}{named arguments giving directories in which RSEM was run.  By default, looks at \code{getConfig()[["RSEM"]]}}
}
\value{
writes a .html
}
\description{
This copies 'adaptorStats.Rmd' to the current working directory
and runs \code{knit2html('adaptorStats.Rmd')} to generate a report on "partition_stats.rds".

Generate a report on mapping/deduplication
}
\details{
You can edit the .Rmd as you see fit to customize the analysis.
}
\section{Functions}{
\itemize{
\item \code{mappingReport}: copies 'qc.Rmd' and generates a report on "deduplicateStats.RData" and on mapping statistics located in various RSEM directories
}}

