% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pear}
\alias{pear}
\title{Run PEAR to assemble paired-end fastq files into one files.}
\usage{
pear(ncores = 10)
}
\arguments{
\item{ncores}{\code{integer} number of cores to use}
}
\description{
Run the pear tool to assemble paired end fastq files into a single output fastq file.
}
\details{
Use this as a preliminary preprocessing step to running MiTCR if you have paired-end data.
}

