% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pipelineReport}
\alias{pipelineReport}
\title{Produce a report listing the tools and packages used by the pipeline.}
\usage{
pipelineReport()
}
\description{
Produce a report listing the tools and packages used by the pipeline.
}
\details{
Helps with reproducibility by producing a report listing the tools and packages used by the pipeline.
}

