% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{QualityControl}
\alias{QualityControl}
\title{Generate QualityControl Matric.}
\usage{
QualityControl(paired = FALSE)
}
\arguments{
\item{paired}{\code{logical} specify whether you have paried reads or not.}
}
\description{
calculate qc statistics: nGenesOn, librarySize, alignment percentage to reference genome, exon mapping rate,
and per base sequence quality output by FASTQC software
}

