% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{qualityRNASeQC}
\alias{qualityRNASeQC}
\title{Run RNASeQC quality control}
\usage{
qualityRNASeQC(picard.path = "/home/jdeng/bin/picard-tools-1.110/",
  ncores = 6, rna_seqc.path = "/home/jdeng/bin/RNA-SeQC_v1.1.7.jar",
  paired = FALSE)
}
\arguments{
\item{picard.path}{\code{character} specifying an \emph{absolute path} path to the PICARD software directory.}

\item{ncores}{\code{integer} number of cores for running in parallel}

\item{rna_seqc.path}{\code{character} specifying an \emph{absolute path} path to the RNASeQC directory.}

\item{paired}{\code{logical} specify whether data is paired. Defaults to FALSE.}
}
\description{
Pre-run Checklist
1. Are the contig names consistent betwen BAM, Refrence and GTF file? 
2. Is the BAM indexed? we index the bam files in the prepareBamFiles
3. Is you reference Indexed? We index the refrence genome before runing the tophat
4. Dose the reference genome have a dict file? we generated the dict file in the createGenomeDict
}
