% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\docType{package}
\name{RNASeqPipeline}
\alias{RNASeqPipeline}
\alias{RNASeqPipeline-package}
\title{Simplify pipelining RNASeq data}
\description{
RNASeqPipeline allows you to standardize the processing of your RNASeq data.
}
\details{
RNASeqPipeline simplifies the process of pipelining the processing of RNASeq
data by helping you organize the raw, intermediate and processed data in a consistent
manner.
}
\author{
Raphael Gottardo and Greg Finak
}

