% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{RSEMAssembleExpressionMatrix}
\alias{RSEMAssembleExpressionMatrix}
\title{Assemble an expression matrix of all results}
\usage{
RSEMAssembleExpressionMatrix(force = FALSE)
}
\arguments{
\item{force}{\code{logical} rerun even if output exists}
}
\description{
Put all the counts from the individual libraries into a single matrix result
}
\details{
Assemble an expression matrix from the individual libraries.
}
