% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RSEMCalculateExpression}
\alias{RSEMCalculateExpression}
\title{Use the RSEM tool to align the reads}
\usage{
RSEMCalculateExpression(ncores = 4)
}
\arguments{
\item{ncores}{\code{integer} specify how many cores to use}
}
\description{
Use the RSEM tool to align reads
}
\details{
Uses RSEM to align reads in FASTQ files against the reference genome.
}

