% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{RSEMSummarizeMapping}
\alias{RSEMSummarizeMapping}
\title{Summarize the number of mapped, unmapped and ambiguous reads from RSEM output}
\usage{
RSEMSummarizeMapping(dir, log = TRUE, plot = TRUE)
}
\arguments{
\item{dir}{Directory containing RSEM cnt files.  If missing, use the RSEM directory from project.}

\item{log}{return log of reads?}

\item{plot}{print a ggplot summary}
}
\value{
returns a data table of # reads per library, invisibly
}
\description{
Summarize the number of mapped, unmapped and ambiguous reads from RSEM output
}

