% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{saveConfig}
\alias{saveConfig}
\title{Save the configuration for a project}
\usage{
saveConfig()
}
\description{
Save the configuration for a project
}
\details{
Save the configuration from a project, stored in the CONFIG directory
}

