% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{sequenceAlignmentTopHat}
\alias{sequenceAlignmentTopHat}
\title{Use the Tophat tool to align reads}
\usage{

  sequenceAlignmentTopHat(path = "/shared/silo_researcher/Gottardo_R/jingyuan_working/iGenome/Mus_musculus/UCSC/mm10",
  parallel_threads = 1, tophat_threads = 6, paired = FALSE,
  nchunks = 10, days_requested = 5, slurm = FALSE,
  slurm_partition = "gottardo_r", ram_per_node = tophat_threads *
  parallel_threads * 1200)
}
\arguments{
\item{path}{\code{character} specifying an \emph{absolute path} path to the iGenome directory.}

\item{parallel_threads}{\code{integer} specify how many parallel processes to spawn}

\item{tophat_threads}{\code{integer} specify how many threads bowtie should use.}

\item{paired}{\code{logical} specify whether you have paried reads or not.}

\item{nchunks}{\code{integer} number of chunks to split the files for a slurm job. Ignored if slurm = FALSE}

\item{days_requested}{\code{integer} number of days requested for the job (when submitting a slurm job). Ignored if slurm = FALSE}

\item{slurm}{\code{logical} if \code{TRUE} job is submitted as a slurm batch job, otherwise it's run on the local machine. Slurm jobs will honour the nchunks and days_requested arguments.}

\item{slurm_partition}{\code{character} the slurm partition to submit to. Ignored if slurm=FALSE}

\item{ram_per_node}{\code{numeric} The number of Mb per node. Ignored if slurm=FALSE. Default of \code{parallel_threads*bowtie_threads*1000}}
}
\description{
Uses Tophat to align reads in FASTQ files against the reference genome hg38 from UCSC database. Optionally you can specify paired end reads. The code assumes 
paired reads have fastq files that differ by one character (i.e. sampleA_read1.fastq, sampleA_read2.fastq) and will perform
matching of paired fastq files based on that assumption using string edit distance. Read 1 is assumed to be upstream
and read 2 is assumed to be downstream.
}
\details{
The number of parallel_threads*tophat_threads should not be more than the number of cores available on your system.
}
\note{
The amount of memory requested should be set to bowtie_threads*parallel_threads*1G as this is the default requested by samtools for sorting. If insufficient memory is requested, the bam files will not be created successfully.
}

