% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{setGenomeReference}
\alias{setGenomeReference}
\title{Set path to reference genome index}
\usage{
setGenomeReference(utils, refName)
}
\arguments{
\item{utils}{\code{character} the path to the directory containing reference genome.}

\item{refName}{\code{character)} name of reference index or genome}
}
\description{
Set path to reference genome index
}
