% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{SRAOverview}
\alias{SRAOverview}
\title{Overview of downloaded files}
\usage{
SRAOverview()
}
\description{
Prints an overview of the SRA files
}
\details{
Generates a table of the downloaded files for an overview to be included
in reports.
}

