% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{sumDuplicates}
\alias{sumDuplicates}
\title{Sum all read counts for multiple instances of a single gene symbol.}
\usage{
sumDuplicates(counts, fDat, cDat = NULL)
}
\arguments{
\item{counts}{\code{matrix} table of read counts or tpms. Rows are genes and columns are
samples}

\item{fDat}{\code{data.frame} feature data - rows map to counts rows}

\item{cDat}{\code{data.frame}. experimental design or phenotype data - rows map to counts columns}
}
\value{
ExpressionSet containing the summed counts data (counts), the feature data which will map to the summed counts data (fDat), and optionally, the experimental data (cDat)
}
\description{
A gene symbol can map to multiple UCSC gene cluster ids so in those
instances all read counts mapped to a single gene symbol are summed.
The summed read counts are returned in an Expression Set along with
the feature data table and optionally an experimental design table.
}

