% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{summarizeDuplication}
\alias{summarizeDuplication}
\title{Summarize duplication statistics from fastqc}
\usage{
summarizeDuplication()
}
\value{
data.table of source data, invisbly
}
\description{
Prints plots of the duplication distribution and % duplication across libraries
}

