% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeFastQC}
\alias{summarizeFastQC}
\title{Summarize the fastQC results}
\usage{
summarizeFastQC()
}
\description{
Summarize the fastqc results
}
\details{
Function to be run after `runFastQC` that will summarize the fastQC results
and generate a plot.
}

