% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{buildReference}
\alias{buildReference}
\title{Build a reference genome}
\usage{
buildReference(path = NULL, gtf_file = NULL, fasta_file = NULL,
  isoformsFile = NULL, name = NULL, doSTAR = TRUE, starPath = NULL,
  threads = 1)
}
\arguments{
\item{path}{\code{character} specifying an \emph{absolute path} path to the Reference_Genome directory.}

\item{gtf_file}{\code{character} the name of the gtf file. Empty by default. If specified the function will look for the named file in the 'path' directory}

\item{fasta_file}{\code{character} the name of the fasta file, must be specified and located in the 'path' directory}

\item{isoformsFile}{\code{character} the name if the known isoforms file (optional). If specified it must be located in the 'path' file.}

\item{name}{\code{character} the prefix of the genome output.}

\item{doSTAR}{\code{logical} logical showing whether to build genome with STAR index.}

\item{starPath}{\code{character}  full path to the 'star' executable.}

\item{threads}{\code{integer} number of threads to use}
}
\description{
Builds a reference genome at `path/Reference_Genome/`
}
\details{
You must specify the Utils path if it is not already defined, and have your genome in a folder titled
`Reference_Genome`. This function will construct the reference genome using RSEM tools.
The command line is the default shown in the documentation.
`rsem-prepare-reference --gtf gtf_file --transcript-to-gene-map knownIsoforms.txt --bowtie2 fasta_file name`
or if doSTAR=TRUE
 `rsem-prepare-reference --gtf gtf_file --transcript-to-gene-map knownIsoforms.txt --star --star-path starPath fasta_file name`
If the gtf_file is not give, then the transcript-to-gene-map option is not used either. A fasta_file and a name must be provided.
}
