% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{downloadSRA}
\alias{downloadSRA}
\title{Download FastQ files based on Immport data and SRA db, unless they already exist.}
\usage{
downloadSRA(GSE_accession = NULL)
}
\arguments{
\item{GSE_accession}{(not implemented)}
}
\description{
Based on the SRA db, download the FastQ files, unless they already exist.
}
\details{
Conditionally download the FastQ files in the SRA db based on Immport tables.
If the files are already present, they will not be downloaded.
}

