% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getConfig}
\alias{getConfig}
\title{Get the configuration for the project from the namespace}
\usage{
getConfig(create = FALSE)
}
\arguments{
\item{create}{/code{logical} If TRUE then this will return empty list that can be used
to create new configuration. If FALSE then standard error checking will be run.}
}
\value{
Returns configuration structure or stops with error if not found
}
\description{
Retrieve the configuration for the project from the namespace. If
configuration not found exit with error.
}

