% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{getExpressionSet}
\alias{getExpressionSet}
\title{Return and expression set of the counts or TPM values}
\usage{
getExpressionSet(which = "counts")
}
\arguments{
\item{which}{\code{character} \code{c("counts","tpm")} specifies what to return.}
}
\description{
Construct and return an expression set of counts or TPM values
}
\details{
Constructs an expression set of counts or tpm values depending on the value of which parameter.
}
