% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deduplicate.R
\name{getUniqueQname}
\alias{getUniqueQname}
\title{Generate a candidate list of *duplicate* reads using UMI/Sequence}
\usage{
getUniqueQname(bamdt, umipattern = "[ACGT]+$", max.edit.dist = 8,
  debug = 1)
}
\arguments{
\item{bamdt}{a data.table with qname, rname, pos and seq}

\item{umipattern}{regular expression matching the UMIs in the qnames}

\item{max.edit.dist}{distance below which we declare two candidate to be duplicates}

\item{debug}{higher numbers result in more verbose output}
}
\value{
list with elements badqname (a character vector of qnames to discard)
}
\description{
The mapping position (POS/rname) is used as heuristic to find reads with
high sequence similarities
}

