% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{loadProject}
\alias{loadProject}
\title{Load an RNASeqPipelineR project}
\usage{
loadProject(project_dir = NULL, name = NULL)
}
\arguments{
\item{project_dir}{\code{character} the path to the project}

\item{name}{\code{character} the name of the project}
}
\description{
Load an RNASeqPipelineR project
}
\details{
Load an RNASeqPipelineR project from disk.
}
