% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{pipelineReport}
\alias{pipelineReport}
\title{Produce a report listing the tools and packages used by the pipeline.}
\usage{
pipelineReport()
}
\description{
Produce a report listing the tools and packages used by the pipeline.
}
\details{
Helps with reproducibility by producing a report listing the tools and packages used by the pipeline.
This needs to be run on the system that produced the results. Output to project directory OUTPUT/pipeline_report.md.
}

