% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{readConfig}
\alias{readConfig}
\title{Read the configuration for a project}
\usage{
readConfig(project = NULL)
}
\arguments{
\item{project}{\code{character} the path to the project folder}
}
\description{
Read the configuration for a project
}
\details{
Read the configuration from a project, stored in the CONFIG directory
}
