% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{RSEMAssembleExpressionMatrix}
\alias{RSEMAssembleExpressionMatrix}
\title{Assemble an expression matrix of all results}
\usage{
RSEMAssembleExpressionMatrix()
}
\description{
Put all the counts from the individual libraries into a single matrix result
}
\details{
Assemble an expression matrix from the individual libraries.
}

