% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{RSEMCalculateExpression}
\alias{RSEMCalculateExpression}
\title{Use the RSEM tool to annotate and quantify the reads}
\usage{
RSEMCalculateExpression(parallel_threads = 6, bowtie_threads = 1,
  paired = FALSE, frag_mean = NULL, frag_sd = NULL, nchunks = 10,
  days_requested = 5, slurm = FALSE, slurm_partition = NULL,
  ram_per_node = bowtie_threads * parallel_threads * 1200,
  fromBAM = FALSE, fromSTAR = FALSE, mail = NULL, force = FALSE)
}
\arguments{
\item{parallel_threads}{\code{integer} specify how many parallel processes to spawn}

\item{bowtie_threads}{\code{integer} specify how many threads bowtie should use.}

\item{paired}{\code{logical} specify whether you have paried reads or not.}

\item{frag_mean}{\code{numeric}}

\item{frag_sd}{\code{numeric} For single ended reads, specifying these might make calculations of effect length more effective. Optional.}

\item{nchunks}{\code{integer} number of chunks to split the files for a slurm job. Ignored if slurm = FALSE}

\item{days_requested}{\code{integer} number of days requested for the job (when submitting a slurm job). Ignored if slurm = FALSE}

\item{slurm}{\code{logical} if \code{TRUE} job is submitted as a slurm batch job, otherwise it's run on the local machine. Slurm jobs will honour the nchunks and days_requested arguments.}

\item{slurm_partition}{\code{character} the slurm partition to submit to. Ignored if slurm=FALSE}

\item{ram_per_node}{\code{numeric} The number of Mb per node. Ignored if slurm=FALSE. Default of \code{parallel_threads*bowtie_threads*1000}}

\item{fromBAM}{\code{logical} if \code{TRUE} then RSEM will attempt to use previously aligned BAM files, in the \code{BAM} directory, rather than fastq files. The file names expected to end with \code{.transcript.bam}. See RSEM documentation for the format these files must obey.}

\item{fromSTAR}{\code{logical} if \code{TRUE} then RSEM will use the
STAR notation for the BAM files}

\item{mail}{email address to send failure message to, if desired.}

\item{force}{If TRUE then quantify all FASTQ/BAM files else only quantify those
files that have not been quantified.}
}
\description{
Use the RSEM tool to annotate and quantify the reads
}
\details{
Uses RSEM to quantify reads in FASTQ files against the reference genome. Optionally you can specify paired end reads. The code assumes
paired reads have fastq files that differ by one character (i.e. sampleA_read1.fastq, sampleA_read2.fastq) and will perform
matching of paired fastq files based on that assumption using string edit distance. Read 1 is assumed to be upstream
and read 2 is assumed to be downstream.
The number of parallel_threads*bowtie_threads should not be more than the number of cores available on your system.
}
\note{
The amount of memory requested should be set to bowtie_threads*parallel_threads*1G 
as this is the default requested by samtools for sorting. If insufficient memory is 
requested, the bam files will not be created successfully.
}
