% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{runFastQC}
\alias{runFastQC}
\title{Perform FastQC quality control}
\usage{
runFastQC(ncores = 8)
}
\arguments{
\item{ncores}{\code{integer} how many threads to use}
}
\description{
Runs fastqc quality control on the FASTQ files
}
\details{
Produces FASTQC reports for each FASTQ file using fastqc
FASTQ input files must have extension \code{fastqc} (capitalization sensitive).
}
