% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runFastQC}
\alias{runFastQC}
\title{Perform FastQC quality control}
\usage{
runFastQC(ncores = 8)
}
\arguments{
\item{ncores}{\code{integer} how many threads to use}
}
\description{
Runs fastqc quality control on the FASTQ files
}
\details{
Produces FASTQC reports for each FASTQ file using fastqc
}

