% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SRAOverview}
\alias{SRAOverview}
\title{Overview of downloaded files}
\usage{
SRAOverview()
}
\description{
Prints an overview of the SRA files
}
\details{
Generates a table of the downloaded files for an overview to be included
in reports.
}

