% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqPipelineR.R
\name{summarizeFastQC}
\alias{summarizeFastQC}
\title{Summarize the fastQC results}
\usage{
summarizeFastQC()
}
\value{
prints a plot and invisibly returns the data.table used to generate the plot
}
\description{
Summarize the fastqc results
}
\details{
Function to be run after `runFastQC` that will summarize the fastQC results
and generate a plot.
}
